+++ src/main/java/totemic_commons/pokefenn/item/equipment/music/ItemRattle.java
@@ -52,14 +52,14 @@ public class ItemRattle extends ItemMusic
                 if(time >= 4 && !player.isSneaking())
                 {
                     time = 0;
-                    TotemUtil.playMusic(world, (int) player.posX, (int) player.posY, (int) player.posZ, musicHandler, 0, 0);
+                    TotemUtil.playMusic(world, player.posX, player.posY, player.posZ, musicHandler, 0, 0);
                     particlesAllAround((WorldServer)world, player.posX, player.posY, player.posZ, false);
                     PacketHandler.sendAround(new PacketSound(x, y, z, "rattle"), player.worldObj.provider.dimensionId, x, y, z);
                 }
                 if(time >= 4 && player.isSneaking())
                 {
                     time = 0;
-                    TotemUtil.playMusicForSelector(player.worldObj, (int) player.posX, (int) player.posY, (int) player.posZ, musicHandler, 0);
+                    TotemUtil.playMusicForSelector(player.worldObj, player.posX, player.posY, player.posZ, musicHandler, 0);
                     particlesAllAround((WorldServer)world, player.posX, player.posY, player.posZ, true);
                     PacketHandler.sendAround(new PacketSound(x, y, z, "rattle"), player.worldObj.provider.dimensionId, x, y, z);
                 }
